requirements.txt
datasets
accelerate
xlsxwriter
pillow
transformers
pandas
numpy
matplotlib
seaborn
tqdm
requests